requirements.txt
groq
tqdm
pandas
python-dotenv
openai==2.16.0
scipy
pyyaml>=6.0
